server/requirements.txt
Django == 1.10.5
psycopg2 == 2.6.2
gevent == 1.2.1
gunicorn == 19.6.0
markdown == 2.6.7
django-filter == 1.0.1
djangorestframework == 3.5.3